Containerization/ContPyTorchMod/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
torch==2.1.0
torchvision==0.16.0
pillow==10.0.0
numpy<2.0